[BE] feat : 개인별 문제 추천 api

See merge request s07-final/S07P31C206!77

+++ backend/Aldy/src/main/java/com/example/demo/domain/dto/member/response/MemberResponseDto.java
@@ -10,7 +10,7 @@ public class MemberResponseDto {
     private String baekjoonId;
     private String nickname;
     private String email;
-    private int tier;
+    private Long tier;
 
     public MemberResponseDto(Member member){
         this.baekjoonId = member.getBaekjoonId();
